Customizable Help Functions
Resolving PR Changes
